% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEW.DAT.R
\docType{data}
\name{LEW.DAT}
\alias{LEW.DAT}
\title{LEW data}
\format{
A data frame with 200 rows and 1 variable:
\describe{
\item{Deflection}{}
}
}
\usage{
data(LEW.DAT)
}
\description{
A test data set used for testing the \code{fourPlot}-function. For more information
see https://www.itl.nist.gov/div898/strd/univ/addinfo/lew.html.
}
\keyword{datasets}
