% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPCC.R
\name{PPCC}
\alias{PPCC}
\title{PPCC Plot}
\usage{

}
\arguments{
\item{data}{A list of data.}

\item{distribution}{The distribution of your data}
}
\value{
plot
}
\description{
The PPCC plot is a graphical technique for identifying the shape parameter
for a distributional family that best describes the data set.
}
\details{

}
\examples{
example here

}
