% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RANDWEIB.DAT.R
\docType{data}
\name{RANDWEIB.DAT}
\alias{RANDWEIB.DAT}
\title{Random Weibull data}
\format{
A data frame with 500 rows and 1 variable:
\describe{
\item{data}{}
}
}
\usage{
data(RANDWEIB.DAT)
}
\description{
A test data set used for testing the \code{PPCC}-function. The data is random data
drawn from a weibull distribution.
}
\keyword{datasets}
