% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPCC.R
\name{weib_qqplot}
\alias{weib_qqplot}
\title{Probability Plot}
\usage{
ProbPlot(data, distribution, shape)
}
\arguments{
\item{data}{A list of data.}

\item{shape}{The known or presumed shape parameter}

\item{distribution}{The distribution that best describes your data.}
}
\value{
A probability plot.
}
\description{
The probability plot is a graphical technique for assessing whether
or not a data set follows a given distribution. It can be used for estimating the
location and scale parameter of the given distribution.
}
\details{
The data are plotted against a theoretical distribution in such a way that
the points should form approximately a straight line. Departures from this straight
line indicate departures from the specified distribution. Estimates of the location
and scale parameters of the distribution are given by the intercept and slope.
}
\examples{
data <- RANDWEIB.DAT$Y
shape <- 2.23
ProbPlot(data=data, distribution="weibull", shape=shape)

}
