% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourPlot.R
\name{fourPlot}
\alias{fourPlot}
\title{fourPlot}
\usage{
fourPlot(data, bins = 11)
}
\arguments{
\item{data}{A list of data values}

\item{bins}{Number of bins to show in the histogram}
}
\value{
A frame with 4 plots
}
\description{
The 4-plot is an EDA technique for testing underlying assumptions. The 4-plot consists of four plots.
A run sequence plot to answer the questions of whether or not the data has fixed location and/or fixed variation.
A lag plot that show you whether or not the data is random.
A histogram to test if the data follows a normal distribution.
A normal probability plot that tests together with the histogram, if the data follows a normal distribution.
}
\examples{
x <- LEW.DAT$Deflection
fourPlot(x)

}
