% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPCC.R
\name{PPCC}
\alias{PPCC}
\title{PPCC Plot}
\usage{
PPCC(data, distribution)
}
\arguments{
\item{data}{A list of data.}

\item{distribution}{The distribution}
}
\value{
plot
}
\description{
PPCC Plot
}
\examples{
example here

}
