% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPCC.R
\name{PPCC}
\alias{PPCC}
\title{PPCC Plot}
\usage{
PPCC(data, distribution)
}
\arguments{
\item{data}{A list of data.}

\item{distribution}{The distribution that best describes your data.}
}
\value{
A Probability Plot Correlation Coefficient Plot.
}
\description{
The \href{https://www.itl.nist.gov/div898/handbook/eda/section3/ppccplot.htm}{PPCC plot} is a graphical technique for identifying the shape parameter
for a distributional family that best describes the data set.
}
\details{
Firstly, you should your data through the \code{tukey_PPCC}-function to find an appropriate distribution
for your data. Then the PPCC plot is used first to find a good value of the shape parameter, and lastly
the probability plot (for example, use the function \code{ProbPlot}) should then be generated to find estimates of the location and scale parameters
and in addition to provide a graphical assessment of the adequacy of the distributional fit.

You can only use data from either a weibull or gamma distribution, so far.
}
\examples{
data <- RANDWEIB.DAT$Y
PPCC(data, "weibull")

}
