% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixPlot.R
\name{sixPlot}
\alias{sixPlot}
\title{Six Plot}
\usage{
sixPlot(X, Y, bins=30)
}
\arguments{
\item{X}{A list of X values.}

\item{Y}{A list of Y values.}

\item{bins}{Number of bins to show in the histogram.}
}
\value{
A frame with 6 plots.
}
\description{
The \href{https://www.itl.nist.gov/div898/handbook/eda/section3/6plot.htm}{6-plot} is a collection of 6 specific graphical techniques
whose purpose is to assess the validity of a Y versus X fit. It returns a list of 6 plots;
\itemize{
\item \href{https://www.itl.nist.gov/div898/handbook/eda/section3/scatterp.htm}{Scatter plot (predicted vs independent variable)}
\item \href{https://www.itl.nist.gov/div898/handbook/eda/section3/scatterp.htm}{Scatter plot (residuals vs independent variable)}
\item \href{https://www.itl.nist.gov/div898/handbook/eda/section3/scatterp.htm}{Scatter plot (residuals vs predicted values)}
\item \href{https://www.itl.nist.gov/div898/handbook/eda/section3/lagplot.htm}{Lag plot of residuals}
\item \href{https://www.itl.nist.gov/div898/handbook/eda/section3/histogra.htm}{Histogram of residuals}
\item \href{https://www.itl.nist.gov/div898/handbook/eda/section3/normprpl.htm}{Normal probability plot of the residuals}
}
}
\details{
The 6-plot outputs a 6plot object with six plots. A method is provided to print them in a frame.
The plots are named scat_plot, res_x_plot, res_pred_plot,lag_res_plot, hist_res_plot and qq_plot.
The model is a linear fit.
\itemize{
\item \code{scat_plot} uses geom_point()
\item \code{res_x_plot} uses geom_point()
\item \code{res_pred_plot} uses geom_point()
\item \code{lag_res_plot} uses geom_point()
\item \code{hist_res_plot} uses geom_histogram()
\item \code{qq_plot} uses geom_qq()
}
}
\examples{
X <- 1:100
Y <- rnorm(100, x * 10, 50)
sixPlot(X, Y)

}
