% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixPlot.R
\name{sixPlot}
\alias{sixPlot}
\title{sixPlot}
\usage{
sixPlot(X, Y)
}
\arguments{
\item{data}{the data}
}
\value{
what it returns
}
\description{
sixPlot
}
\examples{
X <- 1:100
Y <- rnorm(x * 10, 0, 50)
sixPlot(X, Y)

}
