% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmTest.r
\name{tmTest}
\alias{tmTest}
\title{Tietjen-Moore Test for Outliers}
\usage{
tmTest(data, k)
}
\arguments{
\item{data}{A list of univariate data.}

\item{k}{The number of suspected outliers.}
}
\value{
An object that shows the result of the hypothesis.
}
\description{
\href{https://www.itl.nist.gov/div898/handbook/eda/section3/eda35h2.htm}{Tietjen-Moore test} is a function to test for k outliers in both tails.
}
\details{
The Tietjen-Moore test is used to detect multiple outliers in a univariate data set
that follows an approximately normal distribution. If testing for a single outlier, the Tietjen-Moore
test is equivalent to the Grubbs' test. TESTESTESTEST
}
\examples{
x <- c(-1.40, -0.44, -0.30, -0.24, -0.22, -0.13, -0.05, 0.06, 0.10, 0.18, 0.20, 0.39, 0.48, 0.63, 1.01)
k <- 3
tmTest(data=x,k=k)

}
